use upower in battery module
